moved editing data & dash callbacks to own section
